address-profiling-pipeline/mapper/requirements.txt
telicent-ies-tool==1.0.2
telicent-lib==2.0.1
python-dotenv
setuptools
pytest
importlib
